requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
requests>=2.28.0
pydantic>=2.0.0
beautifulsoup4>=4.9.3
notion-client>=1.0.0
pytest>=7.0.0